requirements.txt
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
anthropic>=0.39.0
python-dotenv>=1.2.1
PyGithub>=2.1.1